refactor: extract useTransactionHandlers from App.tsx (phase 4a)
Moves 7 transaction handlers (~570 lines) out of App.tsx into
src/hooks/useTransactionHandlers.ts. The hook receives session,
setError, and setNotice as parameters and returns all handlers.
App.tsx now destructures them from a single hook call.

App.tsx: 1281 → 750 lines. Behavior and all 658 tests unchanged.
